+++ packages/core/examples/simple/main.ts
@@ -4,16 +4,16 @@ import {
 } from 'node:path'
 import { fileURLToPath } from 'node:url'
 
-import { version }  from '../../package.json'
+import { name }     from '../../package.json'
 import { Creatium } from '../../src' // change for `import { Creatium } from 'creatium'`
 
 const currentDir   = join( dirname( fileURLToPath( import.meta.url ) ) )
 const templatesDir = join( currentDir, 'templates' )
-const name         = 'Simple test'
 
 export const creatium = new Creatium( {
 	name,
-	version,
+	version   : '0.2.2', // minor version for display updater notification
+	updater   : true,
 	templates : {
 		js : {
 			input : join( templatesDir, 'js-project' ),