+++ packages/core/src/_shared/up.ts
@@ -2,7 +2,7 @@ import { Updater } from '@clippium/updater'
 
 import { color } from '../utils'
 
-export const updater = ( name: string, version: string ) => {
+export const updater = async ( name: string, version: string ) => {
 
 	const {
 		blue,
@@ -12,20 +12,16 @@ export const updater = ( name: string, version: string ) => {
 		italic,
 	} = color
 
-	async () => {
+	const _updater = new Updater( {
+		version,
+		name,
+	} )
 
-		const _updater = new Updater( {
-			version,
-			name,
-		} )
-		const data     = await _updater.get()
-		if ( !data ) return
+	const data = await _updater.get()
+	if ( !data ) return
 
-		console.log( `📦 ${bold( 'Update available' )} ${gray( data.currentVersion )} → ${green( data.latestVersion )} ${italic( `(${data.type})` )}
-
-Run ${blue( data.packageManager + ' i ' + name )} to update
+	console.log( `║ 📦 ${bold( 'Update available' )} ${gray( data.currentVersion )} → ${green( data.latestVersion )} ${italic( `(${data.type})` )}
+║ Run ${blue( data.packageManager + ' i ' + name )} to update
 		` )
-
-	}
 
 }